Перераспределены пакеты в соответсвии с логическими частями приложения
