--- app/src/main/java/com/practicum/playlistmaker/data/dto/Response.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-open class Response() {
-    var resultCode = 0
-}
