--- app/src/main/java/com/practicum/playlistmaker/data/dto/SearchTracksResponse.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-import com.google.gson.annotations.SerializedName
-
-data class SearchTracksResponse(@SerializedName("results") val tracks: List<TrackDto>): Response()
