--- app/src/main/java/com/practicum/playlistmaker/data/dto/TrackDto.kt
@@ -1,14 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-data class TrackDto(
-    val trackId: Int,
-    val trackName: String,
-    val artistName: String,
-    val trackTimeMillis: Long?,
-    val artworkUrl100: String?,
-    val collectionName: String?,
-    val releaseDate: String?,
-    val primaryGenreName: String?,
-    val country: String?,
-    val previewUrl: String?
-)
