--- app/src/main/java/com/practicum/playlistmaker/data/network/ApiFactory.kt
@@ -1,15 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object ApiFactory {
-    private const val BASE_URL = "https://itunes.apple.com/"
-
-    private val retrofit: Retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    val itunesService: ItunesApi = retrofit.create(ItunesApi::class.java)
-}
