--- app/src/main/java/com/practicum/playlistmaker/data/network/ItunesApi.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import com.practicum.playlistmaker.data.dto.SearchTracksResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ItunesApi {
-    @GET("search?entity=song")
-    fun search(@Query("term") text: String): Call<SearchTracksResponse>
-}
