--- app/src/main/java/com/practicum/playlistmaker/data/storage/HistoryStorage.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.data.storage
-
-import com.practicum.playlistmaker.domain.model.Track
-
-interface HistoryStorage {
-    fun saveTrack(track: Track)
-
-    fun getAllTracks(): List<Track>
-
-    fun clearHistory()
-}
