--- app/src/main/java/com/practicum/playlistmaker/domain/api/HistoryInteractor.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-import com.practicum.playlistmaker.domain.model.Track
-
-interface HistoryInteractor {
-    fun saveTrack(track: Track)
-
-    fun getAllTracks(): List<Track>
-
-    fun clearHistory()
-}
