--- app/src/main/java/com/practicum/playlistmaker/domain/impl/HistoryInteractorImpl.kt
@@ -1,19 +1,0 @@
-package com.practicum.playlistmaker.domain.impl
-
-import com.practicum.playlistmaker.domain.api.HistoryInteractor
-import com.practicum.playlistmaker.domain.api.HistoryRepository
-import com.practicum.playlistmaker.domain.model.Track
-
-class HistoryInteractorImpl(private val repository: HistoryRepository): HistoryInteractor {
-    override fun saveTrack(track: Track) {
-        repository.saveTrack(track)
-    }
-
-    override fun getAllTracks(): List<Track> {
-        return repository.getAllTracks()
-    }
-
-    override fun clearHistory() {
-        repository.clearHistory()
-    }
-}
