+++ app/src/main/java/com/practicum/playlistmaker/player/ui/view_model/PlayerViewModelFactory.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlistmaker.player.ui.view_model
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class PlayerViewModelFactory(private val trackUrl: String) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        return PlayerViewModel(trackUrl) as T
+    }
+}
