+++ app/src/main/java/com/practicum/playlistmaker/search/data/dto/Response.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.search.data.dto
+
+open class Response() {
+    var resultCode = 0
+}
