+++ app/src/main/java/com/practicum/playlistmaker/search/data/dto/SearchTracksResponse.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.search.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class SearchTracksResponse(@SerializedName("results") val tracks: List<TrackDto>): Response()
