+++ app/src/main/java/com/practicum/playlistmaker/search/data/network/ItunesApi.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.search.data.network
+
+import com.practicum.playlistmaker.search.data.dto.SearchTracksResponse
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ItunesApi {
+    @GET("search?entity=song")
+    fun search(@Query("term") text: String): Call<SearchTracksResponse>
+}
