+++ app/src/main/java/com/practicum/playlistmaker/search/data/storage/HistoryStorage.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.search.data.storage
+
+import com.practicum.playlistmaker.search.domain.model.Track
+
+interface HistoryStorage {
+    fun saveTrack(track: Track)
+
+    fun getAllTracks(): List<Track>
+
+    fun clearHistory()
+}
