+++ app/src/main/java/com/practicum/playlistmaker/search/domain/api/HistoryInteractor.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.search.domain.api
+
+import com.practicum.playlistmaker.search.domain.model.Track
+
+interface HistoryInteractor {
+    fun saveTrack(track: Track)
+
+    fun getAllTracks(): List<Track>
+
+    fun clearHistory()
+}
