+++ app/src/main/java/com/practicum/playlistmaker/search/domain/api/HistoryRepository.kt
@@ -1,0 +1,12 @@
+package com.practicum.playlistmaker.search.domain.api
+
+import com.practicum.playlistmaker.search.domain.model.Track
+
+interface HistoryRepository {
+    fun saveTrack(track: Track)
+
+    fun getAllTracks(): List<Track>
+
+    fun clearHistory()
+
+}
