+++ app/src/main/res/layout/activity_library.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.library.LibraryActivity">
+    tools:context=".library.ui.LibraryActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
