+++ app/src/main/res/layout/activity_player.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.player.PlayerActivity">
+    tools:context=".player.ui.activity.PlayerActivity">
 
     <ImageView
         android:id="@+id/arrow_back"
